% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{add.aws.path}
\alias{add.aws.path}
\title{Add aws PATH}
\usage{
add.aws.path(path.to.aws.bin)
}
\arguments{
\item{path.to.aws.bin}{The path to aws bin path.}
}
\description{
Add aws PATH
}
\details{
Currently, this is a OS X and linux based function only; does not work for windows.  This function adds the aws bin to the enviroment variable PATH.  For this function to work correctly, ensure that you set up the config file correctly, which is contained in your \code{.aws} folder (by default the \code{.aws} folder should be set up at \code{~/.aws}).  Specifically, in your config file, you must specify the \code{aws_access_key_id}, \code{aws_secret_access_key} and the default \code{region}. The Default region, \code{aws_access_key_id}, and the \code{aws_secret_access_key}, can be specified by setting the enviroment variables \code{AWS_DEFAULT_REGION}, \code{AWS_ACCESS_KEY_ID}, and \code{AWS_SECRET_ACCESS_KEY}, respectively (e.g. Sys.setenv(AWS_DEFAULT_REGION = "us-east-1"))
}
\examples{
## Not run:
add.aws.path("~/.local/lib/aws/bin")
## End(**Not run**)
}
\keyword{AWS.tools}

