% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cancel.spot.instanceReq}
\alias{cancel.spot.instanceReq}
\title{Cancel Spot Instances Request}
\usage{
cancel.spot.instanceReq(SpotInstanceReqIDs)
}
\arguments{
\item{SpotInstanceReqIDs}{The spot Instances Request IDs to be cancelled.}
}
\description{
Cancel Spot Instances Request
}
\details{
Cancel EC2 Spot instances Requests.
}
\examples{
## Not run:
cancel.spot.instanceReq(SpotInstanceReqIDs = c("sir-9a80844f", "sir-b06d364f"))
## End(**Not run**)
}
\keyword{AWS.tools}

