% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{create.ec2.key}
\alias{create.ec2.key}
\title{Create EC2 Key Pairs}
\usage{
create.ec2.key(key.name, region)
}
\arguments{
\item{key.name}{Name of the key pair that will be created.}

\item{region}{(OPTIONAL) if missing, the default region (defined part of the enviroement variable, or aws config file) is used}
}
\description{
Create EC2 Key Pairs
}
\details{
Create an EC2 key pairs and saves it to the \code{.ssh} folder located in your home directory. EC2 key is need to be able to Launch and connect to an instance on AWS cloud. If no Key pairs is used when launching the instances, the user will not be able to ssh login into his/her instance.
}
\examples{
## Not run:
create.ec2.key("MyEC2Key")
## End(**Not run**)
}
\keyword{AWS.tools}

