% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{delete.ec2.key}
\alias{delete.ec2.key}
\title{Delete EC2 Key Pairs}
\usage{
delete.ec2.key(key.name, region)
}
\arguments{
\item{key.name}{Name of the key pair to delete.}
}
\description{
Delete EC2 Key Pairs
}
\details{
Delete an EC2 key pairs from AWS and removes it from \code{.ssh} folder if it exist.
}
\examples{
## Not run:
delete.ec2.key("MyEC2Key")
## End(**Not run**)
}
\keyword{AWS.tools}

