% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{describe.avail.zone}
\alias{describe.avail.zone}
\title{Describes Availability Zone}
\usage{
describe.avail.zone(region.name)
}
\arguments{
\item{region.name}{Name of the aws region. (OPTIONAL). If missing, the default (aws) region is used}
}
\description{
Describes Availability Zone
}
\details{
Describes Availability Zone in a given region (if specified).
}
\examples{
## Not run:
describe.avail.zone()
describe.avail.zone("us-west-2")
## End(**Not run**)
}
\keyword{AWS.tools}

