% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{describe.instances}
\alias{describe.instances}
\title{Describe instances}
\usage{
describe.instances(region)
}
\arguments{
\item{region}{(OPTIONAL) If missing, default region (defined part the enviromement variable or in the AWS config file)is used}
}
\value{
A matrix where the columns are: LanchTime, InstanceId, State.Name, RootDeviceType, RootDeviceName, KeyName and PublicDns Name if available.
}
\description{
Describe EC2 instances on the AWS cloud.
}
\examples{
describe.instances()
}
\keyword{AWS.tools}

