% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{describe.regions}
\alias{describe.regions}
\title{Describes AWS regions}
\usage{
describe.regions()
}
\description{
Describes AWS regions
}
\details{
Describes AWS regions available so that you can specify the region parameters when creating EC2-key pairs, launching instanaces, requesting spot instances, ... etc. Use the value in the ID columns when specifying regions.
}
\examples{
## Not run:
describe.regions()
## End(**Not run**)
}
\keyword{AWS.tools}

