% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{describe.spot.instanceReq}
\alias{describe.spot.instanceReq}
\title{Describe Spot Instances Request}
\usage{
describe.spot.instanceReq(SpotInstanceReqIDs)
}
\arguments{
\item{SpotInstanceReqIDs}{The spot Instances Request IDs (Optional)}
}
\value{
A matrix whose columns represents: Spot Instance Request ID, the Request Time, the specified Bidding Price, the status update time, the status code, the request state, and if the later is active, the Instance ID
}
\description{
Describe Spot Instances Request
}
\details{
Describes EC2 Spot instances Requests that were placed earlier by the user.
}
\examples{
## Not run:
describe.spot.instanceReq()
## End(**Not run**)
}
\keyword{AWS.tools}

