% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ec2stop.instances}
\alias{ec2stop.instances}
\title{Stop EC2 instance(s) using instance IDs}
\usage{
ec2stop.instances(instance.ids, region)
}
\arguments{
\item{instance.ids}{A character vector specifying instance ids that will be stopped.}

\item{region}{(OPTIONAL) If not specified, default region is used.}
}
\description{
Stops one or more chosen instances currently running on EC2 using instance IDs.
}
\examples{
## Not run:
# Stop one instance with id "i-41df9c6d"
ec2stop.instances(instance.ids = "i-41df9c6d")

# Stop two instances
ec2stop.instances(instance.ids = c("i-41df9c6d","i-41eg1e3d"))
## End(**Not run**)
}
\keyword{AWS.tools}

