% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ec2stop.reservation}
\alias{ec2stop.reservation}
\title{Stop EC2 instances using reservation IDs}
\usage{
ec2stop.reservation(reservation.id, region)
}
\arguments{
\item{reservation.id}{A character string.}

\item{region}{(OPTIONAL) If not specified, default region is used.}
}
\description{
Stop instances on EC2 based on reservation IDs.
}
\examples{
## Not run:
ec2stop.reservation(reservation.id = "r-eaf8ca9b")
## End(**Not run**)
}
\keyword{AWS.tools}

