% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ec2terminate.instances}
\alias{ec2terminate.instances}
\title{Terminate EC2 instances using instance IDs}
\usage{
ec2terminate.instances(instance.ids, region)
}
\arguments{
\item{instance.ids}{a vector of the desired instances id that you wish to stop.}

\item{region}{(OPTIONAL) If not specified, default region is used.}
}
\description{
Terminates one or more chosen instances currently running on EC2.
}
\examples{
## Not run:
 # Terminate one instance with id "i-41df9c6d"
ec2terminate.instances(instance.ids = "i-41df9c6d")
## End(**Not run**)
}
\keyword{AWS.tools}

