% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get.spot.price}
\alias{get.spot.price}
\title{Get Instance Spot Price}
\usage{
get.spot.price(instance.type = "cr1.8xlarge", region)
}
\arguments{
\item{instance.type}{Type of instance desired (eg m1.xlarge).}

\item{region}{(OPTIONAL) If not specified, default region is used.}
}
\description{
Get Instance Spot Price
}
\details{
Get the most recent spot price for a type of instance.
}
\examples{
## Not run:
get.spot.price("cr1.8xlarge")
## End(**Not run**)
}
\keyword{AWS.tools}

