% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{req.spot.instances}
\alias{req.spot.instances}
\title{Request Spot Instances}
\usage{
req.spot.instances(spot.price, ami, instance.count, instance.type, security.grp,
  key.name, region, root_DeleteOnTermination = TRUE, verbose = FALSE)
}
\arguments{
\item{spot.price}{The maximum bidding price the user is willing to pay.}

\item{ami}{The AMI ID the user wishes to use to launch the instances with.}

\item{instance.count}{The number of instances the user wishes to launches}

\item{instance.type}{The type of the machine the user wishes to launches eg. "t1.micro"}

\item{security.grp}{(OPTIONAL). If mssing the default security group is applied to instances that are to be launched.}

\item{key.name}{Name of the key to launch EC2 instances with; in order to connect to them later through ssh.}

\item{region}{The region ID (OPTIONAL). If not specified, the default region is used. See AWS config file, or the enviromemt variable \code{AWS_DEFAULT_REGION} for default value.}

\item{root_DeleteOnTermination}{TRUE(defaults) or FALSE. If set to false, the root volume will NOT be deleted when instance is terminated.}
}
\description{
Request Spot Instances
}
\details{
Request EC2 Spot instances on AWS. Spot instances are cheaper than On Demand Instances.
}
\examples{
## Not run:
req.spot.instances(spot.price = 0.0032, ami = "ami-e84d8480", instance.count = 2, instance.type = "t1.micro", key.name = "MyEC2Key" )
## End(**Not run**)
}
\keyword{AWS.tools}

