% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rootVol.DeleteOnTermination}
\alias{rootVol.DeleteOnTermination}
\title{Make the root volume of an instance Persistent}
\usage{
rootVol.DeleteOnTermination(instance.id)
}
\arguments{
\item{instance.id}{Instance ID for which to make the root volume persistent.}
}
\description{
Make the root volume of an instance Persistent
}
\details{
By default, the root device volume for an AMI backed by Amazon EBS is deleted when the instance terminates. This function changes this default behavior.
}
\examples{
## Not run:
rootVol.DeleteOnTermination("i-5203422c")
## End(**Not run**)
}
\keyword{AWS.tools}

