% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{s3.ls}
\alias{s3.ls}
\title{List Bucket Function}
\usage{
s3.ls(bucket = NULL, recursive = FALSE, filters = NULL)
}
\arguments{
\item{recursive}{Boolean (Default FALSE). Command is performed on allfiles or objects under the specified bucket.}

\item{filters}{character string indicating filters.}

\item{path}{S3 Path/name of bucket. Optional}
}
\description{
List buckets or files within a specified bucket on S3.
}
\details{
filters: a character string See AWS documentaion for the filters usage.
\code{\link{http://docs.aws.amazon.com/cli/latest/reference/s3/index.html#use-of-exclude-and-include-filters}}
}
\examples{
## Not run:
# list all file and directories under mybucket
s3.ls("mybucket")

# list all bucket on S3
s3.ls()

# list all all files under the folder_test which reside inside mybucket
s3.ls("mybucket/folfer_test")
## End(**Not run**)
}
\keyword{AWS.tools}

