% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{s3.mb}
\alias{s3.mb}
\title{Make Bucket Function}
\usage{
s3.mb(bucket)
}
\arguments{
\item{bucket}{Name of bucket to create. REQUIRED}
}
\description{
This function creates a bucket on S3.
}
\examples{
## Not run:
s3.mb("mybucket")
## End(**Not run**)
}
\keyword{AWS.tools}

