% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{s3.mv}
\alias{s3.mv}
\title{Move function}
\usage{
s3.mv(from.path, to.path, recursive = FALSE, ACL = NULL, filters = NULL)
}
\arguments{
\item{from.path}{Path of file to copy from.}

\item{to.path}{Path of file to copy to.}

\item{recursive}{Boolean (Default FALSE). Command is performed on allfiles or objects under the specified bucket.}

\item{ACL}{Access Control list (Optional).}

\item{filters}{A string that define the include and/or exclude filters. See details for more info.}
}
\description{
Move a local file or S3 object to another location locally or in S3.
s3.mv <LocalPath> <S3Path> or <S3Path> <LocalPath> or <S3Path> <S3Path>.
}
\details{
filters: a character string See AWS documentaion for the filters usage.
\code{\link{http://docs.aws.amazon.com/cli/latest/reference/s3/index.html#use-of-exclude-and-include-filters}}.
ACL: Sets the ACl for the object when the command is performed. Allowed values are: \emph{private, public-read, public-read-write, authenticated-read, bucket-owner-read, bucket-owner-full-control and log-delivery-write}.
For more information see \code{\link{http://docs.aws.amazon.com/cli/latest/reference/s3/mv.html}}
\strong{NOTE:}
from.path Path and to.path Path must be prefixed with \preformatted{s3://} if moving from or to S3. See examples below.
}
\examples{
## Not run:
# move local text.txt to my bucket on S3 to a file name test2.txt
s3.mv("test.txt", "s3://mybucket/test2.txt")

# move file from one S3 bucket to another S3 bucket while retaining its original name
s3.mv("s3://mybucket/test.txt", "s3://mybucket2/")

# move ALL files from S3 bucket(mybucket) to current local folder
s3.mv("s3://mybucket", ".", recursive = TRUE)
## End(**Not run**)
}
\keyword{AWS.tools}

