% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{s3.rb}
\alias{s3.rb}
\title{Remove Bucket Function}
\usage{
s3.rb(bucket, force = FALSE)
}
\arguments{
\item{bucket}{Name of bucket to remove. REQUIRED}

\item{force}{Boolean (Default FALSE). Deletes all objects in the bucket including the bucket itself.}
}
\description{
Creates a bucket on S3.
}
\examples{
## Not run:
s3.rb("mybucket")
## End(**Not run**)
}
\keyword{AWS.tools}

