% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{s3.rm}
\alias{s3.rm}
\title{remove Bucket Function}
\usage{
s3.rm(s3.object, is.bucket = FALSE, recursive = FALSE, filters = NULL)
}
\arguments{
\item{recursive}{Boolean (Default FALSE). Command is performed on allfiles or objects under the specified bucket.}

\item{filters}{character string indicating filters.}

\item{S3}{object (bucket of file).}
}
\description{
Delete files within a specified bucket on S3; or entire bucket (if empty or recursive set to TRUE).
}
\details{
filters: a character string See AWS documentaion for the filters usage.
\code{\link{http://docs.aws.amazon.com/cli/latest/reference/s3/index.html#use-of-exclude-and-include-filters}}
}
\examples{
## Not run:
# list all file and directories under mybucket
s3.rm("mybucket", is.bucket = TRUE, recursive = TRUE)

# rm text2.txt which is located under S3 bucket (mybucket)
s3.rm("mybucket/test2.txt")
## End(**Not run**)
}
\keyword{AWS.tools}

