% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{s3.sync}
\alias{s3.sync}
\title{Sync function}
\usage{
s3.sync(from.path, to.path, recursive = FALSE, ACL = NULL, filters = NULL,
  delete = FALSE)
}
\arguments{
\item{from.path}{Path to sync from.}

\item{to.path}{Path to sync to.}

\item{filters}{A string that define the include and/or exclude filters. See details for more info.}

\item{delete}{Boolean (Default FALSE).}
}
\description{
Syncs directories and S3 prefixes.
s3.sync <LocalPath> <S3Path> or <S3Path> <LocalPath> or <S3Path> <S3Path>.
}
\details{
filters: a character string See AWS documentaion for the filters usage.
\code{\link{http://docs.aws.amazon.com/cli/latest/reference/s3/index.html#use-of-exclude-and-include-filters}}.
delete: Files that exist in the destination but not in the source are deleted during sync.
For more information see \code{\link{http://docs.aws.amazon.com/cli/latest/reference/s3/sync.html}}
\strong{NOTE:}
from.path Path and to.path Path must be prefixed with \preformatted{s3://} if moving from or to S3. See examples below.

A file (or s3 object) will be copied (upload/download) if the size of the file (in from.path) is different than the size of the file in the to.path; the last modified time of the local file is newer than the last modified time of file in the to.path; or the file in the to.path does not exist.
}
\examples{
## Not run:
# sync content of local folder to a specified S3 bucket
s3.sync(".", "s3://mybucket")

sync content S3 bucket to local folder
s3.sync("s3://mybucket/test.txt", ".")

# In the following example, the user syncs the bucket mybucket to the local current directory. The local current directory contains the files test.jpg and test2.txt. The bucket mybucket contains the object test.jpg of a different size than the local test.jpg
Because the exclude filters is used, all files matching the pattern existing both in s3 and locally will be excluded from the sync.
s3.sync(".", "s3://mybucket", filters ="--exclude '*.jpg' ")
# will results in uploading: test2.txt to s3://mybucket/test2.txt

# In this example, the user syncs the local current directory to the bucket mybucket. The local current directory contains the files test.txt and another/test2.txt. The bucket mybucket contains the object test1.txt and another/test5.txt:
s3.sync("s3://mybucket/", ".", filters = "--exclude '*another/*' "
# Will results in downloading s3://mybucket/test1.txt to test1.txt
## End(**Not run**)
}
\keyword{AWS.tools}

