% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{startCluster}
\alias{startCluster}
\title{Start an EC2 cluster}
\usage{
startCluster(ami, instance.count, instance.type, security.grp, key.name, region,
  verbose = FALSE, root_DeleteOnTermination = TRUE)
}
\arguments{
\item{ami}{The image-id for the instance to use.  This will be specific to your account (log into AWS to see what image-id's are available to you).}

\item{instance.count}{An integer number of instances to start.}

\item{instance.type}{Type of instance (eg. t1.micro).}

\item{security.grp}{(OPTIONAL) Security group to be applied to the cluster. If you don't specify a security group when launching an instance, Amazon EC2 uses the default security group.}

\item{key.name}{The Name of the key to launch the EC2 instances with in order to connect to them later through ssh.}

\item{region}{(OPTIONAL) If missing, the default region is used.}

\item{verbose}{TRUE/FALSE Print to console the command used.}

\item{root_DeleteOnTermination}{TRUE/FALSE. If set to TRUE (Default); the Root Device Volume is deleted when the instance Terminates/Fails.}
}
\description{
Create an EC2 cluster on the AWS cloud.
}
\examples{
## Not run:
# starts a cluster of 2 nodes of type t1.micro
myCluster <- startCluster(ami = "ami-80778be8", instance.count = 2, instance.type = "t1.micro")
## End(**Not run**)
}
\keyword{AWS.tools}

