% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{terminateCluster}
\alias{terminateCluster}
\title{Terminate an EC2 cluster}
\usage{
terminateCluster(cluster, region)
}
\arguments{
\item{cluster}{An \code{ec2.cluster} object created with startCluster function.}

\item{region}{(OPTIONAL) If not specified, default region is used.}
}
\description{
Terminate a specified cluster on EC2.
}
\examples{
## Not run:
myCluster <- startCluster(ami = "ami-80778be8", instance.count = 2, instance.type = "t1.micro")
terminateCluster(cluster = myCluster)
## End(**Not run**)
}
\keyword{AWS.tools}

