% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{upload.key}
\alias{upload.key}
\title{Upload Public Key}
\usage{
upload.key(path_public_key, key_name, region)
}
\arguments{
\item{path_public_key}{The path to the public key file (e.g "~/.ssh/id_rsa.pub")}

\item{key_name}{Name of the key pair that will be created on AWS}

\item{region}{(OPTIONAL) The name of the region to upload the key too}
}
\description{
Upload Public Key
}
\details{
Upload a public key from the user machine to the AWS cloud to be used to launch/connect to an EC2 instance.
}
\examples{
## Not run:
upload.key("~/.ssh/id_rsa.pub", "awsKey")
## End(**Not run**)
}
\keyword{AWS.tools}

